requirements.txt
# --- API ---
fastapi[standard]
fastapi
uvicorn[standard]
requests

# --- AWS ---
boto3
botocore

# --- Hugging Face ---
transformers

#--- Local model --- 
sentencepiece
accelerate

# --- Embeddings locales ---
sentence-transformers
torch

# --- Vector search ---
faiss-cpu

# --- Data utils ---
numpy
pandas

# --- Misc ---
python-dotenv
pydantic
# -- Gradio --
gradio
